+++ libnopegl/src/node_block.c
@@ -136,13 +136,13 @@ void ngli_node_block_extend_usage(struct ngl_node *node, int usage)
     s->usage |= usage;
 }
 
-int ngli_node_block_get_cpu_size(struct ngl_node *node)
+size_t ngli_node_block_get_cpu_size(struct ngl_node *node)
 {
     struct block_info *s = node->priv_data;
     return s->data_size;
 }
 
-int ngli_node_block_get_gpu_size(struct ngl_node *node)
+size_t ngli_node_block_get_gpu_size(struct ngl_node *node)
 {
     struct block_info *s = node->priv_data;
     return s->data_size;
